// src/modules/database-driver/connector/grpc/gen/GrpcEnums_pb.ts
export enum GrpcOrderDirection {
  /**
   * Ascending order.
   *
   * @generated from enum value: ASC = 0;
   */
  ASC = 0,

  /**
   * Descending order.
   *
   * @generated from enum value: DESC = 1;
   */
  DESC = 1,
}
// ...
export enum GrpcOrderBehaviour {
  /**
   * Null values are sorted before non-null values.
   *
   * @generated from enum value: NULLS_FIRST = 0;
   */
  NULLS_FIRST = 0,

  /**
   * Null values are sorted after non-null values.
   *
   * @generated from enum value: NULLS_LAST = 1;
   */
  NULLS_LAST = 1,
}
// ...
export enum GrpcCardinality {
  /**
   * No cardinality specified.
   *
   * @generated from enum value: NOT_SPECIFIED = 0;
   */
  NOT_SPECIFIED = 0,

  /**
   * Relation may be missing completely, but if it exists - there is never more than single relation of this type.
   *
   * @generated from enum value: ZERO_OR_ONE = 1;
   */
  ZERO_OR_ONE = 1,

  /**
   * There is always single relation of this type.
   *
   * @generated from enum value: EXACTLY_ONE = 2;
   */
  EXACTLY_ONE = 2,

  /**
   * Relation may be missing completely, but there may be also one or more relations of this type.
   *
   * @generated from enum value: ZERO_OR_MORE = 3;
   */
  ZERO_OR_MORE = 3,

  /**
   * There is always at least one relation of this type, but there may be also more than one.
   *
   * @generated from enum value: ONE_OR_MORE = 4;
   */
  ONE_OR_MORE = 4,
}
// ...
export enum GrpcEvolutionMode {
  /**
   * When first entity is inserted in the collection - primary key generation will automatically adapt whether
   * this first entity has the primary key already present or not. If it is present all other entities are expected
   * to have also primary key provided from external systems, if not primary keys will be always automatically
   * generated by Evita and may never be provided from outside.
   *
   * @generated from enum value: ADAPT_PRIMARY_KEY_GENERATION = 0;
   */
  ADAPT_PRIMARY_KEY_GENERATION = 0,

  /**
   * When new attribute is encountered, it is silently added to a collection schema as non-filterable, non-sortable,
   * non-unique attribute of the type that was used for the first value. Localizability of the attribute is determined
   * by the fact whether the first value encountered was associated with any localization.
   *
   * @generated from enum value: ADDING_ATTRIBUTES = 1;
   */
  ADDING_ATTRIBUTES = 1,

  /**
   * When new associated data is encountered, it is silently added to a collection schema. Localizability of
   * the associated data is determined by the fact whether the first value encountered was associated with any
   * localization.
   *
   * @generated from enum value: ADDING_ASSOCIATED_DATA = 2;
   */
  ADDING_ASSOCIATED_DATA = 2,

  /**
   * When new reference type is encountered, it is silently added. It's linked to other evitaDB entity when
   * the entity type of a reference or its group respectively matches by
   * a type of a different entity reference any existing entity collection. Reference is not indexed, nor
   * hierarchy indexed by default and this must be changed by updating the schema.
   *
   * @generated from enum value: ADDING_REFERENCES = 3;
   */
  ADDING_REFERENCES = 3,

  /**
   * When entity has no prices and new price is encountered, it is silently added. That means that entity type
   * starts supporting prices when first price has been added.
   *
   * @generated from enum value: ADDING_PRICES = 4;
   */
  ADDING_PRICES = 4,

  /**
   * When new localization is encountered, it is silently added.
   *
   * @generated from enum value: ADDING_LOCALES = 5;
   */
  ADDING_LOCALES = 5,

  /**
   * When new currency is encountered in prices, it is silently added.
   *
   * @generated from enum value: ADDING_CURRENCIES = 6;
   */
  ADDING_CURRENCIES = 6,

  /**
   * When hierarchy placement for the entity is encountered, it is silently added.
   *
   * @generated from enum value: ADDING_HIERARCHY = 7;
   */
  ADDING_HIERARCHY = 7,
}
// ...
export enum GrpcAttributeSchemaType {
  /**
   * attribute schema is GlobalAttributeSchemaContract
   *
   * @generated from enum value: GLOBAL = 0;
   */
  GLOBAL = 0,

  /**
   * attribute schema is EntityAttributeSchemaContract
   *
   * @generated from enum value: ENTITY = 1;
   */
  ENTITY = 1,

  /**
   * attribute schema is AttributeSchemaContract
   *
   * @generated from enum value: REFERENCE = 2;
   */
  REFERENCE = 2,
}